Add test for empty title validation in RegisterExpense
Added a new `using` directive for `CashFlow.Exception` in the
`RegisterExpenseValidatorTests.cs` file. Introduced a new test
method `ErrorTitleEmpyt` in the `RegisterExpenseValidatorTests`
class. This test method creates an instance of `RegisterExpenseValidator`,
builds a request using `RequestRegisterExpenseBuilder`, sets the
`Title` property of the request to an empty string, validates the
request using the validator, asserts that the validation result is
invalid, and asserts that the validation errors contain a single
error with the message `ResourceErrorMessages.TITLE_REQUIRED`.